condense even more, only need to change a single line for new datasets
